task-72: move MembersPage to pages
