+++ src/js/view/pages/MembersPage/index.ts
@@ -1,0 +1,1 @@
+export * from "./MembersPage";
